# packages/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "bmad-todo API",
    "version": "0.0.0"
  },
  "components": {
    "schemas": {}
  },
  "paths": {
    "/todos": {
      "get": {
        "summary": "List todos",
        "tags": [
          "todos"
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "headers": {
              "x-request-id": {
                "schema": {
                  "required": true,
                  "type": "string"
                },
                "description": "Request correlation identifier"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "todos"
                  ],
                  "properties": {
                    "todos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "text",
                          "completed",
                          "createdAt",
                          "updatedAt"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "text": {
                            "type": "string"
                          },
                          "completed": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "deletedAt": {
                            "anyOf": [
                              {
                                "type": "string",
                                "format": "date-time"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default Response",
            "headers": {
              "x-request-id": {
                "schema": {
                  "required": true,
                  "type": "string"
                },
                "description": "Request correlation identifier"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create todo",
        "tags": [
          "todos"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "additionalProperties": false,
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "pattern": ".*\\S.*"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "headers": {
              "x-request-id": {
                "schema": {
                  "required": true,
                  "type": "string"
                },
                "description": "Request correlation identifier"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "text",
                    "completed",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    },
                    "completed": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "deletedAt": {
                      "anyOf": [
                        {
                          "type": "string",
                          "format": "date-time"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "headers": {
              "x-request-id": {
                "schema": {
                  "required": true,
                  "type": "string"
                },
                "description": "Request correlation identifier"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "Default Response",
            "headers": {
              "x-request-id": {
                "schema": {
                  "required": true,
                  "type": "string"
                },
                "description": "Request correlation identifier"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}